Updated posts, users, and admin endpoints
