+++ app/routers/users.py
@@ -22,48 +22,7 @@ async def get_users(db: db_dependency):
     return get_user_model
 
 
-@router.get(
-    "/get_user_by_id", status_code=status.HTTP_200_OK, response_model=UserResponse
-)
-async def get_user_by_id_or_accound_id(
-    db: db_dependency,
-    user_id: Optional[int] = Query(None),
-    account_id: Optional[UUID] = Query(None),
-):
-    if not user_id and not account_id:
-        raise HTTPException(
-            status_code=status.HTTP_400_BAD_REQUEST,
-            detail="At least one ID (user_id or account_id) must be provided",
-        )
 
-    user_query = db.query(Users)
-
-    if user_id:
-        user_query = user_query.filter(Users.id == user_id)
-    if account_id:
-        user_query = user_query.filter(Users.account_id == account_id)
-
-    user = user_query.first()
-
-    if not user:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND, detail="User not found"
-        )
-
-    user_response = UserResponse(
-        id=user.id,
-        account_id=user.account_id,
-        username=user.username,
-        email=user.email,
-        bio=user.bio,
-        avatar=user.avatar,
-        user_type=user.role,
-        is_active=user.is_active,
-        created_at=user.created_at,
-        last_login=user.last_login,
-    )
-
-    return user_response
 
 
 @router.put("/update_user", status_code=status.HTTP_200_OK)
@@ -116,10 +75,8 @@ async def update_user_email(
     return {"detail": "User email updated successfully"}
 
 
-@router.delete("/delete_user/{user_id}", status_code=status.HTTP_200_OK)
-async def delete_user(
-    user: user_dependency, db: db_dependency, user_id: int = Path(ge=0)
-):
+@router.delete("/delete_user", status_code=status.HTTP_200_OK)
+async def delete_user(user: user_dependency, db: db_dependency):
 
     if user is None:
         raise HTTPException(
@@ -130,9 +87,5 @@ async def delete_user(
 
     db.query(Users).filter(Users.id == user_id).delete()
     db.commit()
-    if Users.id != user_id:
-        raise HTTPException(
-            status_code=status.HTTP_400_BAD_REQUEST,
-            detail="User not authorized",
-        )
+   
     return {"detail": "User deleted successfully"}
